# src/server/api/swagger.json
{
  "swagger": "2.0",
  "produces": [
    "application/json"
  ],
  "info": {
    "title": "conversationdiscovery",
    "description": "An API",
    "version": "1.0"
  },
  "basePath": "/conversationdiscovery",
  "schemes": [
    "http",
    "https"
  ],
  "parameters": {
    "clientId": {
      "name": "clientId",
      "in": "path",
      "description": "A client id, e.g. \"247inc\".",
      "required": true,
      "type": "string"
    },
    "applicationId": {
      "name": "applicationId",
      "in": "path",
      "description": "An app id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
...
    "accountId": {
      "name": "accountId",
      "in": "path",
      "description": "An account id, e.g. \"ReferenceBot\".",
      "required": true,
      "type": "string"
    },
...
    "runId": {
      "name": "runId",
      "in": "path",
      "description": "Run id of a run",
      "required": true,
      "type": "string",
      "format": "uuid"
    },
    "clusterId": {
      "name": "clusterId",
      "in": "path",
      "description": "Cluster identifier",
      "required": true,
      "type": "string"
    },
...
    "projectId": {
      "name": "projectId",
      "in": "path",
      "description": "id of the project",
      "required": true,
      "type": "string"
    },
...
    "interfaceId": {
      "name": "interfaceId",
      "in": "path",
      "description": "interfaceId",
      "required": true,
      "type": "string"
    },
...
  },
  "responses": {
    "BadRequest": {
      "description": "The request was invalid. For example, the value of the initial parameter was invalid",
      "schema": {
        "type": "string"
      }
    },
...
    "Unauthorized": {
      "description": "Missing or incorrect authentication credentials. This may also returned in other undefined circumstances.Do not retry with the same input",
      "schema": {
        "type": "string"
      }
    },
...
    "InternalServerError": {
      "description": "Internal server error",
      "schema": {
        "type": "string"
      }
    },
...
  },
  "paths": {
...
    "/clients/{clientId}/interfaces": {
      "get": {
        "summary": "get list of all interfaces for answers",
        "description": "get list of all interfaces",
        "x-swagger-router-controller": "answers",
        "operationId": "getInterfaceList",
        "tags": [
          "answer"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          }
        ],
        "responses": {
          "200": {
            "description": "Getting Interface List successful"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
		    }
      }
    },
    "/clients/{clientId}/accounts/{accountId}/applications/{applicationId}/projects/{projectId}/runs/{runId}/clusters/{clusterId}/interfaceId/{interfaceId}/addToAnswers": {
      "post": {
        "summary": "create faq for answers api",
        "description": "create faq for answers api",
        "x-swagger-router-controller": "answers",
        "operationId": "createFAQ",
        "tags": [
          "faqs"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/interfaceId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/clusterId"
          }
        ],
        "responses": {
          "200": {
            "description": "created successfully"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "patch": {
        "summary": "update response for answers",
        "description": "update response for answers",
        "x-swagger-router-controller": "answers",
        "operationId": "updateFAQ",
        "tags": [
          "faqs"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/interfaceId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/clusterId"
          }
        ],
        "responses": {
          "200": {
            "description": "created successfully"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      },
      "get": {
        "summary": "Get list of all FAQs for answers",
        "description": "Get list of all FAQs for answers from Candidate Folder",
        "x-swagger-router-controller": "answers",
        "operationId": "getFAQs",
        "tags": [
          "faqs"
        ],
        "parameters": [
          {
            "$ref": "#/parameters/clientId"
          },
          {
            "$ref": "#/parameters/interfaceId"
          },
          {
            "$ref": "#/parameters/projectId"
          },
          {
            "$ref": "#/parameters/runId"
          },
          {
            "$ref": "#/parameters/applicationId"
          },
          {
            "$ref": "#/parameters/accountId"
          },
          {
            "$ref": "#/parameters/clusterId"
          }
        ],
        "responses": {
          "200": {
            "description": "Getting FAQs List successful"
          },
          "400": {
            "$ref": "#/responses/BadRequest"
          },
          "401": {
            "$ref": "#/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/responses/InternalServerError"
          }
        }
      }
     }
  },
...
}